Move class functions to class_manager
