+++ src/omop_etl_wrapper/model/table_manager/__init__.py
@@ -1,1 +1,2 @@
 from .vocabulary_manager import VocabManager
+from .concept_class_manager import ClassManager
